+++ src/NeverNull/Combinators/ContainsExt.cs
@@ -2,16 +2,19 @@
 using System.Collections.Generic;
 
 namespace NeverNull.Combinators {
+    /// <summary>
+    ///     Provides extension methods to check if a specified <see cref="Option{T}" /> contains a specific value.
+    /// </summary>
     public static class ContainsExt {
         /// <summary>
-        ///     Returns a value that indicates if this option contains the desired value.
+        ///     Returns a value that indicates if the specified <paramref name="option" /> contains the
+        ///     <paramref name="desiredValue" />.
         /// </summary>
         /// <typeparam name="T"></typeparam>
         /// <param name="option"></param>
         /// <param name="desiredValue"></param>
         /// <param name="comparer"></param>
         /// <returns></returns>
-        /// <exception cref="ArgumentNullException"></exception>
         public static bool Contains<T>(this Option<T> option, T desiredValue, IEqualityComparer<T> comparer = null) {
             var c = comparer ?? EqualityComparer<T>.Default;
 
@@ -19,19 +22,23 @@ namespace NeverNull.Combinators {
         }
 
         /// <summary>
-        ///     Returns a value that indicates if this option contains the desired value.
+        ///     Returns a value that indicates if the specified <paramref name="option" /> contains the
+        ///     <paramref name="desiredValue" />.
         /// </summary>
         /// <typeparam name="T"></typeparam>
         /// <param name="option"></param>
         /// <param name="desiredValue"></param>
         /// <param name="compare"></param>
         /// <returns></returns>
-        /// <exception cref="ArgumentNullException"></exception>
+        /// <exception cref="ArgumentNullException">
+        ///     <paramref name="compare" /> is <see langword="null" />
+        /// </exception>
         public static bool Contains<T>(this Option<T> option, T desiredValue, Func<T, T, bool> compare) {
             compare.ThrowIfNull(nameof(compare));
 
-            T t;
-            return option.TryGet(out t) && compare(t, desiredValue);
+            return option.Match(
+                None: () => false,
+                Some: x => compare(x, desiredValue));
         }
     }
 }